Breaks up the route file into several smaller modules
still some work to be done as far as break up but it is begun.
Much smaller and easier to reason about now that the file is shorter
and the modules will now each be testable which they were not
before.
